moved ICE to z80core, changed cpmsim to use it
1. Moved z80sim/srcsim/disas.c and most of z80sim/srcsim/simctl.c to
   z80core/simdis.c and z80core/simice.c respectively.
2. Changed z80sim, mosteksim and cpmsim to use the simdis.c & simice.c
   See comment before function ice_cmd_loop() in simice.c for usage
   info.
3. Some trailing white space in Makefiles removed.
